CAUSEWAY-3390: makes design of AuditTrailEntryRepository consistent with ...
... secman's ApplicationUserRepository
